feat: export as module
