+++ package.json
@@ -7,9 +7,7 @@
   "main": "dist/index.js",
   "module": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "repository": {
     "type": "git",
     "url": "https://github.com/tsshogi/kanna.git"
@@ -23,7 +21,7 @@
     }
   },
   "scripts": {
-    "build": "bun build src/index.ts --sourcemap=external --minify --outdir dist",
+    "build": "bun build src/index.ts --sourcemap=external --target bun --minify --outdir dist",
     "prebuild": "rimraf dist",
     "postbuild": "bunx tsc && tsc-alias",
     "lint": "bunx @biomejs/biome lint --write src",
