+++ src/enums/species.ts
