+++ src/models/coop/coop_history.dto.ts
@@ -9,8 +9,8 @@ import { RawId } from '@/models/common/raw_id.dto'
 import { S3URL } from '@/models/common/s3_url.dto'
 import { WeaponInfoMainHash } from '@/models/common/weapon_hash.dto'
 import { z } from 'zod'
+import type { ResourceQuery } from '../resource.interface.dto'
 import { CoopSchedule } from './coop_schedule.dto'
-import type { ResourceQuery } from './resource.interface'
 
 const HistoryDetail = z.object({
   id: CoopHistoryDetailId,
