+++ src/models/coop/coop_history_detail.dto.ts
@@ -4,9 +4,9 @@ import { CoopEvent } from '@/enums/coop/coop_event'
 import { CoopGrade } from '@/enums/coop/coop_grade'
 import { CoopRule } from '@/enums/coop/coop_rule'
 import { CoopWaterLevel } from '@/enums/coop/coop_water_level'
+import { Species } from '@/enums/species'
 import { WeaponInfoMain } from '@/enums/weapon/main'
 import { WeaponInfoSpecial } from '@/enums/weapon/special'
-import { Species } from '@/enums/weapon/species'
 import { CoopData } from '@/models/common/coop_data.dto'
 import { CoopGradeModel } from '@/models/common/coop_grade.dto'
 import { CoopHistoryDetailId } from '@/models/common/coop_history_detail_id.dto'
@@ -18,8 +18,8 @@ import { RawId, RawInt } from '@/models/common/raw_id.dto'
 import { S3URL } from '@/models/common/s3_url.dto'
 import { WeaponInfoMainHash } from '@/models/common/weapon_hash.dto'
 import { z } from 'zod'
+import type { ResourceQuery } from '../resource.interface.dto'
 import { CoopSchedule } from './coop_schedule.dto'
-import type { ResourceQuery } from './resource.interface'
 
 const BossResultModel = z
   .object({
