+++ src/models/coop/coop_schedule.dto.ts
