+++ src/models/resource.interface.dto.ts
