+++ src/tests/results/index.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, test } from 'bun:test'
-import { readFileSync, readdirSync, writeFileSync } from 'node:fs'
+import { readFileSync, readdirSync } from 'node:fs'
 import path from 'node:path'
-import { CoopResultQuery } from '@/models/coop_result.dto'
+import { CoopResultQuery } from '@/models/coop/coop_result.dto'
 
 describe('CoopResultQuery', () => {
   test('Validity', () => {
