+++ src/tests/schedules/index.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from 'bun:test'
 import { readFileSync, readdirSync } from 'node:fs'
 import path from 'node:path'
-import { CoopScheduleQuery } from '@/models/coop_schedule.dto'
+import { CoopScheduleQuery } from '@/models/coop/coop_schedule.dto'
 
 describe('CoopScheduleQuery', () => {
   test('Validity', () => {
