+++ tsconfig.json
@@ -18,5 +18,5 @@
     "types": ["@types/bun", "@types/node"]
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", ".devcontainer"]
+  "exclude": ["node_modules", ".devcontainer", "dist/**/*", "src/tests/**/*"]
 }
